Adjusted files for merge
