+++ dcs-dtc/Models/AH64/AH64Configuration.cs
@@ -101,12 +101,7 @@ namespace DTC.Models.AH64
                 var lon = pos.Element("Longitude")?.Value;
                 var dLat = double.Parse(lat.Replace('.', ','));
                 var dLon = double.Parse(lon.Replace('.', ','));
-                //if (
-                //    !double.TryParse(lat, out var dLat) ||
-                //    !double.TryParse(lon, out var dLon))
-                //{
-                //    continue;
-                //}
+
                 float.TryParse(pos.Element("Altitude")?.Value.Replace('.', ','), out var elevation);
                 var coord = new Coordinate(dLat, dLon);
                 lat = $"{(dLat > 0 ? 'N' : 'S')} {coord.Latitude.Degrees:00}.{coord.Latitude.DecimalMinute:00.000}";
@@ -116,9 +111,6 @@ namespace DTC.Models.AH64
 
                 previous.Waypoints.Waypoints.Add(new Waypoint(
                     i,"WP","WP","", coord.MGRS.LongZone.ToString() + coord.MGRS.LatZone.ToString()+coord.MGRS.Digraph.ToString()+" "+coord.MGRS.Easting.ToString("00000").Substring(0,4) + " " + coord.MGRS.Northing.ToString("00000").Substring(0, 4),
-                    //string.IsNullOrEmpty(name) ? "" : name,
-                    //string.IsNullOrEmpty(lat) ? "N 00.00.000" : lat,
-                    //string.IsNullOrEmpty(lon) ? "E 000.00.000" : lon,
                     (int)Math.Floor(elevation * feetPerMeter)
                 ));
