+++ dcs-dtc/Models/AH64/Waypoints/WaypointSystem.cs
@@ -63,10 +63,5 @@ namespace DTC.Models.AH64.Waypoints
                 wpt.Sequence = i + 1;
             }
         }
-
-        public int Count()
-        {
-            return Waypoints.Count;
-        }
     }
 }
