--- dcs-dtc/Models/AH64/temp
@@ -1,1 +1,0 @@
-vnbvbn
