+++ dcs-dtc/Models/F16/Waypoints/WaypointSystem.cs
@@ -63,9 +63,5 @@ namespace DTC.Models.F16.Waypoints
 				wpt.Sequence = i + 1;
 			}
 		}
-		public int Count()
-        {
-            return Waypoints.Count;
-        }
     }
 }
